api/requirements.txt
torch
torchvision
pafy==0.5.5
youtube_dl==2020.12.2
opencv-python~=4.5.5.62
websockets
ffmpeg-python
dlib